Port of experimental regrid bilinear rectilinear.
