+++ lib/iris/tests/results/analysis/linear/both_circular.cml
