+++ lib/iris/tests/results/analysis/linear/both_circular.data.0.npy
