+++ lib/iris/tests/results/analysis/linear/circular_grid.cml
