+++ lib/iris/tests/results/analysis/linear/circular_grid.data.0.npy
