+++ lib/iris/tests/results/analysis/linear/circular_src.cml
