+++ lib/iris/tests/results/analysis/linear/circular_src.data.0.npy
