+++ lib/iris/tests/results/analysis/linear/masked_altitude.cml
