+++ lib/iris/tests/results/analysis/linear/masked_altitude.data.0.npy
