+++ lib/iris/tests/results/analysis/linear/no_overlap.data.0.npy
