+++ lib/iris/tests/results/analysis/linear/non_circular.data.0.npy
