+++ lib/iris/tests/results/analysis/linear/partial_overlap.data.0.npy
