+++ lib/iris/tests/results/analysis/linear/subset.data.0.npy
