+++ lib/iris/tests/results/analysis/linear/subset_anon.cml
