+++ lib/iris/tests/results/analysis/linear/subset_anon.data.0.npy
