+++ lib/iris/tests/results/analysis/linear/subset_masked_1.cml
