+++ lib/iris/tests/results/analysis/linear/subset_masked_1.data.0.npy
