+++ lib/iris/tests/results/analysis/linear/subset_masked_2.data.0.npy
