+++ lib/iris/tests/results/visual_tests/test_linear.Test___call____visual.test_osgb_to_latlon.0.png
