+++ lib/iris/tests/results/visual_tests/test_linear.Test___call____visual.test_subsample.0.png
